requirements.txt
google-adk
fastmcp
